don't read server.cfg as client in webhook module
The client will never have a -config line set and the webhook module
will therefore read the default example file (if it exists) and
erroneously set the hostname (to e.g. "EXAMPLE NAME"). We shouldn't
even read this file and the client can already get the server hostname
without it.